move shared IndigoFacade
